Issue #6 Fix serialization issues
Use jackson annotations to control serialization for some enums. Without annotations
the uppercase constant name is used, which is not compatible with the API.

Use jackson annotation for boolean 'isSecondFactorOnly' field. Without explicit
property naming through the annotation jackson defaults to removing the 'is' prefix
for boolean fields during serialization which is not compatible with the API.

Fix  non API compliant enum constant field in 'UserVerification' enum.